Move audio related code to a separate location
Decoupling the code to make it much easier to modify in the future.
