+++ src/fheroes2/castle/castle_dialog.cpp
@@ -24,8 +24,8 @@
 #include <cassert>
 #include <string>
 
-#include "agg.h"
 #include "agg_image.h"
+#include "audio_manager.h"
 #include "army_bar.h"
 #include "castle.h"
 #include "castle_building_info.h"
@@ -258,7 +258,7 @@ Castle::CastleDialogReturnValue Castle::OpenDialog( const bool readOnly, const b
 
     const fheroes2::StandardWindow background( fheroes2::Display::DEFAULT_WIDTH, fheroes2::Display::DEFAULT_HEIGHT );
 
-    AGG::PlayMusic( MUS::FromRace( race ), true, true );
+    fheroes2::PlayMusic( MUS::FromRace( race ), true, true );
 
     int alphaHero = 255;
     CastleDialog::FadeBuilding fadeBuilding;
@@ -279,7 +279,7 @@ Castle::CastleDialogReturnValue Castle::OpenDialog( const bool readOnly, const b
         }
 
         if ( build != BUILD_NOTHING ) {
-            AGG::PlaySound( M82::BUILDTWN );
+            fheroes2::PlaySound( M82::BUILDTWN );
             fadeBuilding.StartFadeBuilding( build );
         }
 
@@ -288,7 +288,7 @@ Castle::CastleDialogReturnValue Castle::OpenDialog( const bool readOnly, const b
 
             generateHeroImage( surfaceHero, heroes );
 
-            AGG::PlaySound( M82::BUILDTWN );
+            fheroes2::PlaySound( M82::BUILDTWN );
             alphaHero = 0;
         }
     }
@@ -621,7 +621,7 @@ Castle::CastleDialogReturnValue Castle::OpenDialog( const bool readOnly, const b
                                 Dialog::Message( _( "Town" ), _( "This town may not be upgraded to a castle." ), Font::BIG, Dialog::OK );
                             }
                             else if ( Dialog::OK == DialogBuyCastle( true ) ) {
-                                AGG::PlaySound( M82::BUILDTWN );
+                                fheroes2::PlaySound( M82::BUILDTWN );
                                 fadeBuilding.StartFadeBuilding( BUILD_CASTLE );
                             }
                             break;
@@ -643,7 +643,7 @@ Castle::CastleDialogReturnValue Castle::OpenDialog( const bool readOnly, const b
                             }
 
                             if ( build != BUILD_NOTHING ) {
-                                AGG::PlaySound( M82::BUILDTWN );
+                                fheroes2::PlaySound( M82::BUILDTWN );
                                 fadeBuilding.StartFadeBuilding( build );
                             }
 
@@ -657,7 +657,7 @@ Castle::CastleDialogReturnValue Castle::OpenDialog( const bool readOnly, const b
                                     topArmyBar.Redraw();
                                     RedrawIcons( *this, CastleHeroes( nullptr, heroes.Guard() ), cur_pt );
                                 }
-                                AGG::PlaySound( M82::BUILDTWN );
+                                fheroes2::PlaySound( M82::BUILDTWN );
 
                                 bottomArmyBar.SetArmy( &heroes.Guest()->GetArmy() );
                                 generateHeroImage( surfaceHero, heroes );
