+++ src/fheroes2/game/game_over.cpp
@@ -22,7 +22,7 @@
  ***************************************************************************/
 
 #include "game_over.h"
-#include "agg.h"
+#include "audio_manager.h"
 #include "campaign_savedata.h"
 #include "castle.h"
 #include "dialog.h"
@@ -96,7 +96,7 @@ namespace
             }
         }
 
-        AGG::PlayMusic( MUS::VICTORY, false );
+        fheroes2::PlayMusic( MUS::VICTORY, false );
 
         if ( !body.empty() )
             Dialog::Message( "", body, Font::BIG, Dialog::OK );
@@ -152,7 +152,7 @@ namespace
             break;
         }
 
-        AGG::PlayMusic( MUS::LOSTGAME, false );
+        fheroes2::PlayMusic( MUS::LOSTGAME, false );
 
         if ( !body.empty() )
             Dialog::Message( "", body, Font::BIG, Dialog::OK );
@@ -326,12 +326,12 @@ fheroes2::GameMode GameOver::Result::LocalCheckGameOver()
                     res = fheroes2::GameMode::COMPLETE_CAMPAIGN_SCENARIO;
                 }
                 else {
-                    AGG::ResetAudio();
+                    fheroes2::ResetAudio();
                     Video::ShowVideo( "WIN.SMK", Video::VideoAction::WAIT_FOR_USER_INPUT );
                     // TODO : Implement function that displays the last frame of win.smk and
                     // a dialog for name entry. AGG:PlayMusic is run here in order to start playing
                     // before displaying the high score.
-                    AGG::PlayMusic( MUS::VICTORY, true, true );
+                    fheroes2::PlayMusic( MUS::VICTORY, true, true );
 
                     res = fheroes2::GameMode::HIGHSCORES_STANDARD;
 
@@ -374,7 +374,7 @@ fheroes2::GameMode GameOver::Result::LocalCheckGameOver()
                         DialogLoss( result );
                     }
 
-                    AGG::ResetAudio();
+                    fheroes2::ResetAudio();
                     Video::ShowVideo( "LOSE.SMK", Video::VideoAction::LOOP_VIDEO );
 
                     res = fheroes2::GameMode::MAIN_MENU;
@@ -385,7 +385,7 @@ fheroes2::GameMode GameOver::Result::LocalCheckGameOver()
     else {
         // There are no active human-controlled players left, game over
         if ( activeHumanColors == 0 ) {
-            AGG::ResetAudio();
+            fheroes2::ResetAudio();
             Video::ShowVideo( "LOSE.SMK", Video::VideoAction::LOOP_VIDEO );
 
             res = fheroes2::GameMode::MAIN_MENU;
@@ -426,7 +426,7 @@ fheroes2::GameMode GameOver::Result::LocalCheckGameOver()
             if ( multiplayerResult & GameOver::WINS ) {
                 DialogWins( multiplayerResult );
 
-                AGG::ResetAudio();
+                fheroes2::ResetAudio();
                 Video::ShowVideo( "WIN.SMK", Video::VideoAction::WAIT_FOR_USER_INPUT );
 
                 res = fheroes2::GameMode::MAIN_MENU;
@@ -434,7 +434,7 @@ fheroes2::GameMode GameOver::Result::LocalCheckGameOver()
             else if ( multiplayerResult & GameOver::LOSS ) {
                 DialogLoss( multiplayerResult );
 
-                AGG::ResetAudio();
+                fheroes2::ResetAudio();
                 Video::ShowVideo( "LOSE.SMK", Video::VideoAction::LOOP_VIDEO );
 
                 res = fheroes2::GameMode::MAIN_MENU;
