+++ src/fheroes2/gui/ui_language.cpp
@@ -73,7 +73,7 @@ namespace fheroes2
 
     SupportedLanguage getResourceLanguage()
     {
-        const std::vector<uint8_t> & data = ::AGG::ReadChunk( ICN::GetString( ICN::FONT ) );
+        const std::vector<uint8_t> & data = ::AGG::getDataFromAggFile( ICN::GetString( ICN::FONT ) );
         if ( data.empty() ) {
             // How is it possible to run the game without a font?
             assert( 0 );
